♻️ refactor(Words): Admin 권한 확인 로직 위치를 Service -> Controller 레이어로 변경

+++ src/main/java/com/feelory/feelory_backend/words/controller/WordsController.java
@@ -3,6 +3,7 @@ package com.feelory.feelory_backend.words.controller;
 
 import com.feelory.feelory_backend.global.api.ApiResponse;
 import com.feelory.feelory_backend.global.api.SuccessCode;
+import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.words.model.*;
 import com.feelory.feelory_backend.words.service.WordsService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -19,6 +20,7 @@ import org.springframework.web.bind.annotation.*;
 public class WordsController {
 
     private final WordsService wordsService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Operation(
             summary = "모든 단어 목록",
@@ -37,6 +39,8 @@ public class WordsController {
     )
     @PostMapping("")
     public ApiResponse<WordCreateResponse> registerWord(@Valid @RequestBody WordCreateRequest request) {
+        jwtTokenProvider.checkAdmin();
+
         WordCreateResponse response = wordsService.registerWord(request);
 
         return ApiResponse.success(response, SuccessCode.REGISTER_WORD_SUCCESS);
@@ -48,6 +52,8 @@ public class WordsController {
     )
     @PatchMapping("")
     public ApiResponse<WordUpdateResponse> patchWord(@Valid @RequestBody WordUpdateRequest request) {
+        jwtTokenProvider.checkAdmin();
+
         WordUpdateResponse response = wordsService.modifyWord(request);
 
         return ApiResponse.success(response, SuccessCode.UPDATE_WORD_SUCCESS);
@@ -59,6 +65,8 @@ public class WordsController {
     )
     @DeleteMapping("")
     public ApiResponse<WordDeleteResponse> deleteWord(@Valid WordDeleteRequest request) {
+        jwtTokenProvider.checkAdmin();
+
         WordDeleteResponse response = wordsService.removeWord(request);
 
         return ApiResponse.success(response, SuccessCode.DELETE_WORD_SUCCESS);
